填写日期: 2019-02-28 今日完成工作:  1.完成车架号扫描功能模块，实现通过扫描车架号，识别车架号，查询车辆信息，  包括品牌，型号，配置，排量，生产年份，指导价。  2.更新版本号。  3.当前版本号 versionCode 20   versionName "1.4.3"

+++ mvp/src/main/java/com/juner/mvp/bean/Technician.java
@@ -21,14 +21,14 @@ public class Technician extends SelectedBean implements Parcelable {
     String email;
     String mobile;
     String roleName;
-    List<Long> roleList;
+    List<Integer> roleList;
     String userSn;
 
-    public List<Long> getRoleList() {
+    public List<Integer> getRoleList() {
         return roleList;
     }
 
-    public void setRoleList(List<Long> roleList) {
+    public void setRoleList(List<Integer> roleList) {
         this.roleList = roleList;
     }
 
@@ -170,7 +170,7 @@ public class Technician extends SelectedBean implements Parcelable {
         this.email = in.readString();
         this.mobile = in.readString();
         this.roleName = in.readString();
-        this.roleList = new ArrayList<Long>();
+        this.roleList = new ArrayList<Integer>();
         in.readList(this.roleList, Long.class.getClassLoader());
         this.userSn = in.readString();
     }
